+++ backend/src/main/java/com/backend/backend/station/StationController.java
@@ -1,5 +1,8 @@
 package com.backend.backend.station;
 
+import jakarta.servlet.http.HttpSession;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 import java.sql.SQLException;
@@ -7,7 +10,10 @@ import java.util.List;
 
 @RestController
 @RequestMapping("api/station")
-@CrossOrigin(origins = "http://localhost:4200") //for development
+@CrossOrigin(
+        origins = "http://localhost:4200",
+        allowCredentials = "true"
+) //for development
 public class StationController {
     private final StationService stationService;
 
@@ -16,13 +22,21 @@ public class StationController {
     }
 
     @GetMapping("/get-list")
-    public List<Station> getStation() throws Exception {
+    public Object getStation(HttpSession session) throws Exception {
+        if (session.getAttribute("userId") == null) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+        }
         return stationService.getStations();
     }
 
     @PostMapping("/new")
-    public Station createStation(@RequestBody Station station) throws Exception {
+    public Object createStation(@RequestBody Station station, HttpSession session) throws Exception {
         try {
+            if("USER".equals(session.getAttribute("role"))) {
+                return ResponseEntity
+                        .status(HttpStatus.FORBIDDEN)
+                        .build();
+            }
             return stationService.createNewStation(station);
         } catch (Exception e) {
             throw e;
@@ -30,8 +44,13 @@ public class StationController {
     }
 
     @PutMapping("/update/{id}")
-    public Station updateStation(@PathVariable String id,  @RequestBody Station station) throws Exception {
+    public Object updateStation(@PathVariable String id,  @RequestBody Station station, HttpSession session) throws Exception {
         try {
+            if("USER".equals(session.getAttribute("role"))) {
+                return ResponseEntity
+                        .status(HttpStatus.FORBIDDEN)
+                        .build();
+            }
             return stationService.updateStation(Integer.parseInt(id), station);
         } catch (Exception e) {
             throw new RuntimeException(e);
@@ -39,11 +58,18 @@ public class StationController {
     }
 
     @DeleteMapping("/delete/{id}")
-    public void deleteStation(@PathVariable Integer id) throws Exception {
+    public Object deleteStation(@PathVariable Integer id, HttpSession session) throws Exception {
         try {
+            if("USER".equals(session.getAttribute("role"))) {
+                return ResponseEntity
+                        .status(HttpStatus.FORBIDDEN)
+                        .build();
+            }
             stationService.deleteStation(id);
         } catch (SQLException e) {
             throw new RuntimeException(e);
         }
+
+        return null;
     }
 }
